hook:02change-repo-sync-type-to-git.sh: check git before sync

+++ hooks/02change-repo-sync-type-to-git.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-if [ ! -d "${WORKDIR}"/squashfs/var/db/repos/gentoo/.git ];then
+if ( chroot "${WORKDIR}"/squashfs which git ) && [ ! -d "${WORKDIR}"/squashfs/var/db/repos/gentoo/.git ];then
     crun rm -rf /var/db/repos/gentoo
     crun emerge --sync
 fi
